CORE-2439: isAllDay flag in event model should be saved in formData. r=jianmin
In modify-event-view, isAllDay flag is a key flag to identify all-day event,
but we should not change the flag before calling persistEvent, since we may
cancel the view and have no chance to reset the event model.
we should use allday element checkbox property to decide the logic of
those elements related to all-day.